Code review returns

+++ server/src/metadata/field-metadata/utils/convert-field-metadata-to-column-action.util.ts
@@ -125,7 +125,7 @@ export function convertFieldMetadataToColumnActions(
         {
           action: WorkspaceMigrationColumnActionType.CREATE,
           columnName: fieldMetadata.targetColumnMap.amountMicros,
-          columnType: 'integer',
+          columnType: 'numeric',
           defaultValue: serializeDefaultValue(defaultValue?.amountMicros),
         },
         {
